Feat/menampilkan pengajuan surat

+++ situ/src/main/java/apap/tugasakhir/situ/repository/PengajuanSuratDb.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface PengajuanSuratDb extends JpaRepository<PengajuanSuratModel, Long> {
-    PengajuanSuratModel getById(Long id);
+    PengajuanSuratModel getById(Integer id);
 }
